Refactor some code to image util
